==> scripts/0_packages.R <==
# install recent versions of R and RStudio

# install recent versions of the packages WITH THE COMMANDS BELOW

# if you get a message about a missing or irremovable package, install that package and try again

install.packages("terra")
install.packages("raster")
install.packages("modEvA", repos = "http://R-Forge.R-project.org")
install.packages("fuzzySim", repos = "http://R-Forge.R-project.org")
install.packages("gam")
install.packages("maxnet")
